# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BrandFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Schedules ────────────────────────────────────────────────────────────
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List scheduled posts
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Schedule a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── AI Generation ────────────────────────────────────────────────────────
...
  /team/invite:
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Removed

  # ─── Dashboard summaries (wow endpoints) ──────────────────────────────────
...
components:
  schemas:
...
    Post:
      type: object
      required: [id, content, platform, status, createdAt]
      properties:
        id:
          type: number
        content:
          type: string
        caption:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        platform:
          type: string
          enum: [twitter, linkedin, instagram, facebook]
        status:
          type: string
          enum: [draft, scheduled, published, failed]
        tone:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        engagementScore:
          type: ["number", "null"]
        likes:
          type: ["number", "null"]
        shares:
          type: ["number", "null"]
        comments:
          type: ["number", "null"]
        reach:
          type: ["number", "null"]
        createdAt:
          type: string
        scheduledAt:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
    PostInput:
      type: object
      required: [content, platform]
      properties:
        content:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        platform:
          type: string
          enum: [twitter, linkedin, instagram, facebook]
        tone:
          type: string
        imageUrl:
          type: string
        scheduledAt:
          type: string
    PostUpdate:
      type: object
      properties:
        content:
          type: string
        caption:
          type: string
        hashtags:
          type: string
        platform:
          type: string
        tone:
          type: string
        status:
          type: string
        imageUrl:
          type: string
        scheduledAt:
          type: string

    # ─── Schedule ────────────────────────────────────────────────────────────
    Schedule:
      type: object
      required: [id, postId, scheduledAt, platform, status, createdAt]
      properties:
        id:
          type: number
        postId:
          type: number
        scheduledAt:
          type: string
        platform:
          type: string
        status:
          type: string
          enum: [pending, sent, failed, cancelled]
        createdAt:
          type: string
        post:
          $ref: "#/components/schemas/Post"
    ScheduleInput:
      type: object
      required: [postId, scheduledAt, platform]
      properties:
        postId:
          type: number
        scheduledAt:
          type: string
        platform:
          type: string
    ScheduleUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
        status:
          type: string

    # ─── AI ──────────────────────────────────────────────────────────────────
...
    TeamMember:
      type: object
      required: [id, email, role, status, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, editor, viewer]
        status:
          type: string
          enum: [active, invited, inactive]
        createdAt:
          type: string
    TeamInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]

    # ─── Dashboard ───────────────────────────────────────────────────────────
...